A whole mess of QOL changes, starting to get the flow for building and installing car files

+++ pkg/ops/car_unpack.go
@@ -26,11 +26,13 @@ var (
 type CarUnpack struct {
 	Info      data.CarInfo
 	Signature []byte
+	Sum       []byte
 }
 
 const (
 	CarInfoJson    = ".car-info.json"
 	SignatureEntry = "~signature"
+	MetadataOnly   = "#fakedir"
 )
 
 func (r *CarUnpack) Install(in io.Reader, dir string) error {
@@ -44,6 +46,8 @@ func (r *CarUnpack) Install(in io.Reader, dir string) error {
 	tr := tar.NewReader(gz)
 
 	dh, _ := blake2b.New256(nil)
+
+	metadataOnly := dir == MetadataOnly
 
 	var sig []byte
 
@@ -82,13 +86,17 @@ top:
 			continue top
 		}
 
-		path := filepath.Join(dir, hdr.Name)
-		dir := filepath.Dir(path)
+		var path, sdir string
 
-		if _, err := os.Stat(dir); err != nil {
-			err = os.MkdirAll(dir, 0755)
-			if err != nil {
-				return err
+		if !metadataOnly {
+			path = filepath.Join(dir, hdr.Name)
+			sdir = filepath.Dir(path)
+
+			if _, err := os.Stat(sdir); err != nil {
+				err = os.MkdirAll(sdir, 0755)
+				if err != nil {
+					return err
+				}
 			}
 		}
 
@@ -97,17 +105,21 @@ top:
 			fmt.Fprintf(dh, hdr.Name)
 			dh.Write([]byte{0})
 
-			mode := hdr.FileInfo().Mode()
-			f, err := os.OpenFile(path, os.O_CREATE|os.O_TRUNC|os.O_WRONLY, mode)
-			if err != nil {
-				return err
-			}
+			if metadataOnly {
+				io.Copy(dh, tr)
+			} else {
+				mode := hdr.FileInfo().Mode()
+				f, err := os.OpenFile(path, os.O_CREATE|os.O_TRUNC|os.O_WRONLY, mode)
+				if err != nil {
+					return err
+				}
 
-			io.Copy(io.MultiWriter(dh, f), tr)
+				io.Copy(io.MultiWriter(dh, f), tr)
 
-			err = f.Close()
-			if err != nil {
-				return err
+				err = f.Close()
+				if err != nil {
+					return err
+				}
 			}
 		case tar.TypeSymlink:
 			fmt.Fprintf(dh, hdr.Name)
@@ -115,9 +127,14 @@ top:
 			fmt.Fprintf(dh, hdr.Linkname)
 			dh.Write([]byte{0})
 
-			err = os.Symlink(filepath.Join(path, hdr.Linkname), path)
-			if err != nil {
-				return err
+			// We normalize all links to be relative in pack, so we just
+			// use them as is here.
+
+			if !metadataOnly {
+				err = os.Symlink(hdr.Linkname, path)
+				if err != nil {
+					return err
+				}
 			}
 		}
 	}
@@ -125,22 +142,29 @@ top:
 	dh.Write(infoData)
 
 	if r.Info.Signer == "" || len(sig) == 0 {
-		os.RemoveAll(dir)
+		if !metadataOnly {
+			os.RemoveAll(dir)
+		}
 		return ErrNoSignature
 	}
 
 	signer, err := base58.Decode(r.Info.Signer)
 	if err != nil {
-		os.RemoveAll(dir)
+		if !metadataOnly {
+			os.RemoveAll(dir)
+		}
 		return err
 	}
 
 	if !ed25519.Verify(ed25519.PublicKey(signer), dh.Sum(nil), sig) {
-		os.RemoveAll(dir)
+		if !metadataOnly {
+			os.RemoveAll(dir)
+		}
 		return ErrInvalidSignature
 	}
 
 	r.Signature = sig
+	r.Sum = h.Sum(nil)
 
 	return nil
 }
